requirements.txt
openai
python-dotenv
motor
bcrypt
jose
bson